--- ShapeCrawler.IntegrationTests/UnitTest1.cs
@@ -1,9 +1,0 @@
-namespace ShapeCrawler.IntegrationTests;
-
-public class UnitTest1
-{
-    [Fact]
-    public void Test1()
-    {
-    }
-}
